fix mismatching dtype

+++ models/matcher.py
@@ -62,7 +62,7 @@ class HungarianMatcher(nn.Module):
         out_bbox = outputs["pred_boxes"].flatten(0, 1)  # [batch_size * num_queries, 4]
 
         # Also concat the target labels and boxes
-        tgt_ids = torch.cat([v["labels"] for v in targets])
+        tgt_ids = torch.cat([v["labels"] for v in targets]).int()
         tgt_bbox = torch.cat([v["boxes"] for v in targets])
 
         # Compute the classification cost.
